+++ tests/StreamWrapper/ResourceStreamWrapperTest.php
@@ -14,7 +14,6 @@ namespace Puli\Repository\Tests\StreamWrapper;
 use PHPUnit_Framework_Assert;
 use PHPUnit_Framework_TestCase;
 use Puli\Repository\Api\ResourceNotFoundException;
-use Puli\Repository\Api\ResourceRepository;
 use Puli\Repository\Resource\FileResource;
 use Puli\Repository\StreamWrapper\ResourceStreamWrapper;
 use Puli\Repository\Tests\Resource\TestDirectory;